Rename surround-thumbnails to just surround.
