+++ src/components/timeline.ts
@@ -1,10 +1,10 @@
 // TODO: When a media viewer is first loaded the selected child won't work (because the underlying carousel has not yet rendered)
-// TODO: rename surround to surround basic and this file to surround?
 // TODO: thumbnails in drawers don't work.
 // TODO: delete segments if not in summary? is this actually necessary? could it create gaps in data? better off stopping access via summary?
 // TODO: support filtering created dataviews by recordings or mediatype (so storage )
 // TODO: dataview refresh instead of rewriteitem?
 // TODO: Make minitimeline configurable in the editor
+// TODO: Is it really useful to select the children in the main timeline view on range change?
 
 import { CSSResultGroup, html, LitElement, TemplateResult, unsafeCSS } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
@@ -12,12 +12,12 @@ import timelineStyle from '../scss/timeline.scss';
 import { CameraConfig, ExtendedHomeAssistant, TimelineConfig } from '../types';
 import { TimelineDataManager } from '../utils/timeline-data-manager';
 import { View } from '../view';
-import './surround-thumbnails.js';
+import './surround.js';
 import './timeline-core.js';
 
 // This file is kept separate from timeline-core.ts to avoid a circular dependency:
 //   FrigateCardTimeline ->
-//   FrigateCardSurroundThumbnails ->
+//   FrigateCardSurround ->
 //   FrigateCardTimelineCore 
 
 @customElement('frigate-card-timeline')
@@ -46,7 +46,7 @@ export class FrigateCardTimeline extends LitElement {
       return html``;
     }
 
-    return html` <frigate-card-surround-thumbnails
+    return html` <frigate-card-surround
       .hass=${this.hass}
       .view=${this.view}
       .thumbnailConfig=${this.timelineConfig.controls.thumbnails}
@@ -62,7 +62,7 @@ export class FrigateCardTimeline extends LitElement {
         .timelineDataManager=${this.timelineDataManager}
       >
       </frigate-card-timeline-core>
-    </frigate-card-surround-thumbnails>`;
+    </frigate-card-surround>`;
   }
 
   /**
