--- src/scss/surround-thumbnails.scss
@@ -1,5 +1,0 @@
-:host {
-  width: 100%;
-  height: 100%;
-  display: block;
-}
